# backend/generated/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CellInfo defines model for CellInfo.
type CellInfo struct {
	CellColumn         int32  `json:"cellColumn"`
	CellRow            int32  `json:"cellRow"`
	CropItem           string `json:"cropItem"`
	DateHarvestPlanted string `json:"dateHarvestPlanted"`
	DatePlanted        string `json:"datePlanted"`
	GrowthStage        int32  `json:"growthStage"`
	Memo               string `json:"memo"`
	PoorGrowthReason   int32  `json:"poorGrowthReason"`
	Status             int32  `json:"status"`
	Variety            string `json:"variety"`
}

// FarmInfo defines model for FarmInfo.
type FarmInfo struct {
	FarmName string             `json:"farmName"`
	FarmUuid openapi_types.UUID `json:"farmUuid"`
}

// FarmInfoDetail defines model for FarmInfoDetail.
type FarmInfoDetail struct {
	FarmLength   string  `json:"farmLength"`
	FarmName     string  `json:"farmName"`
	FarmSeasons  []int   `json:"farmSeasons"`
	FarmWidth    string  `json:"farmWidth"`
	FarmYear     int     `json:"farmYear"`
	FurrowNumber *int    `json:"furrowNumber,omitempty"`
	FurrowWidth  *string `json:"furrowWidth,omitempty"`
}

// FarmPageBasicInfo defines model for FarmPageBasicInfo.
type FarmPageBasicInfo struct {
	FarmManageUuid openapi_types.UUID `json:"farmManageUuid"`
	FarmName       string             `json:"farmName"`
	FarmSeasons    []int32            `json:"farmSeasons"`
	FarmYear       int32              `json:"farmYear"`
}

// FurrowBasicInfo defines model for FurrowBasicInfo.
type FurrowBasicInfo struct {
	FurrowNumber int32 `json:"furrowNumber"`
}

// FurrowCellInfo defines model for FurrowCellInfo.
type FurrowCellInfo struct {
	CellInfoArray   []CellInfo `json:"CellInfoArray"`
	FurrowLength    int32      `json:"furrowLength"`
	FurrowWidth     int32      `json:"furrowWidth"`
	MinPlantSpacing int32      `json:"minPlantSpacing"`
	Rows            int32      `json:"rows"`
}

// Item defines model for Item.
type Item struct {
	Id    *int64  `json:"id,omitempty"`
	Title *string `json:"title,omitempty"`
}

// LoginOrder defines model for LoginOrder.
type LoginOrder struct {
	GroupId  *string `json:"groupId,omitempty"`
	Password *string `json:"password,omitempty"`
	UserId   *string `json:"userId,omitempty"`
}

// LoginUser defines model for LoginUser.
type LoginUser struct {
	GroupUuid *openapi_types.UUID `json:"groupUuid,omitempty"`
	UserUuid  *openapi_types.UUID `json:"userUuid,omitempty"`
}

// PostGroupsGroupUuidManageFarmsJSONRequestBody defines body for PostGroupsGroupUuidManageFarms for application/json ContentType.
type PostGroupsGroupUuidManageFarmsJSONRequestBody = FarmInfoDetail

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = LoginOrder

// PutManageFarmsFarmUuidFarmManageUuidJSONRequestBody defines body for PutManageFarmsFarmUuidFarmManageUuid for application/json ContentType.
type PutManageFarmsFarmUuidFarmManageUuidJSONRequestBody = FarmInfoDetail

// PostManageFarmsFarmUuidFarmManageUuidRowIdJSONRequestBody defines body for PostManageFarmsFarmUuidFarmManageUuidRowId for application/json ContentType.
type PostManageFarmsFarmUuidFarmManageUuidRowIdJSONRequestBody = FurrowCellInfo

// PutManageFarmsFarmUuidFarmManageUuidRowIdJSONRequestBody defines body for PutManageFarmsFarmUuidFarmManageUuidRowId for application/json ContentType.
type PutManageFarmsFarmUuidFarmManageUuidRowIdJSONRequestBody = FurrowCellInfo

// PostManageFarmsFarmUuidFarmManageUuidRowIdCellJSONRequestBody defines body for PostManageFarmsFarmUuidFarmManageUuidRowIdCell for application/json ContentType.
type PostManageFarmsFarmUuidFarmManageUuidRowIdCellJSONRequestBody = CellInfo

// PutManageFarmsFarmUuidFarmManageUuidRowIdCellJSONRequestBody defines body for PutManageFarmsFarmUuidFarmManageUuidRowIdCell for application/json ContentType.
type PutManageFarmsFarmUuidFarmManageUuidRowIdCellJSONRequestBody = CellInfo

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get Farms by GroupUUID
	// (GET /groups/{groupUuid}/manageFarms)
	GetGroupsGroupUuidManageFarms(c *gin.Context, groupUuid openapi_types.UUID)
	// Register New Farm
	// (POST /groups/{groupUuid}/manageFarms)
	PostGroupsGroupUuidManageFarms(c *gin.Context, groupUuid openapi_types.UUID)
	// Order to login
	// (POST /login)
	PostLogin(c *gin.Context)
	// Get Farm Page Info
	// (GET /manageFarms/{farmUuid})
	GetManageFarmsFarmUuid(c *gin.Context, farmUuid openapi_types.UUID)
	// Get furrow Info
	// (GET /manageFarms/{farmUuid}/{farmManageUuid})
	GetManageFarmsFarmUuidFarmManageUuid(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID)
	// Update Farm FarmInfo
	// (PUT /manageFarms/{farmUuid}/{farmManageUuid})
	PutManageFarmsFarmUuidFarmManageUuid(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID)
	// Get Furrow Cell Info
	// (GET /manageFarms/{farmUuid}/{farmManageUuid}/{rowId})
	GetManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID, rowId int32)
	// Create Furrow Cell Info
	// (POST /manageFarms/{farmUuid}/{farmManageUuid}/{rowId})
	PostManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID, rowId int32)
	// Update Furrow Cell Info
	// (PUT /manageFarms/{farmUuid}/{farmManageUuid}/{rowId})
	PutManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID, rowId int32)
	// Create Cell Info
	// (POST /manageFarms/{farmUuid}/{farmManageUuid}/{rowId}/cell)
	PostManageFarmsFarmUuidFarmManageUuidRowIdCell(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID, rowId int32)
	// Update Cell Info
	// (PUT /manageFarms/{farmUuid}/{farmManageUuid}/{rowId}/cell)
	PutManageFarmsFarmUuidFarmManageUuidRowIdCell(c *gin.Context, farmUuid openapi_types.UUID, farmManageUuid openapi_types.UUID, rowId int32)
	// Get all users
	// (GET /managePlant)
	GetManagePlant(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetGroupsGroupUuidManageFarms operation middleware
func (siw *ServerInterfaceWrapper) GetGroupsGroupUuidManageFarms(c *gin.Context) {

	var err error

	// ------------- Path parameter "groupUuid" -------------
	var groupUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "groupUuid", c.Param("groupUuid"), &groupUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter groupUuid: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetGroupsGroupUuidManageFarms(c, groupUuid)
}

// PostGroupsGroupUuidManageFarms operation middleware
func (siw *ServerInterfaceWrapper) PostGroupsGroupUuidManageFarms(c *gin.Context) {

	var err error

	// ------------- Path parameter "groupUuid" -------------
	var groupUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "groupUuid", c.Param("groupUuid"), &groupUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter groupUuid: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostGroupsGroupUuidManageFarms(c, groupUuid)
}

// PostLogin operation middleware
func (siw *ServerInterfaceWrapper) PostLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostLogin(c)
}

// GetManageFarmsFarmUuid operation middleware
func (siw *ServerInterfaceWrapper) GetManageFarmsFarmUuid(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetManageFarmsFarmUuid(c, farmUuid)
}

// GetManageFarmsFarmUuidFarmManageUuid operation middleware
func (siw *ServerInterfaceWrapper) GetManageFarmsFarmUuidFarmManageUuid(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetManageFarmsFarmUuidFarmManageUuid(c, farmUuid, farmManageUuid)
}

// PutManageFarmsFarmUuidFarmManageUuid operation middleware
func (siw *ServerInterfaceWrapper) PutManageFarmsFarmUuidFarmManageUuid(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutManageFarmsFarmUuidFarmManageUuid(c, farmUuid, farmManageUuid)
}

// GetManageFarmsFarmUuidFarmManageUuidRowId operation middleware
func (siw *ServerInterfaceWrapper) GetManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "rowId" -------------
	var rowId int32

	err = runtime.BindStyledParameterWithOptions("simple", "rowId", c.Param("rowId"), &rowId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter rowId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetManageFarmsFarmUuidFarmManageUuidRowId(c, farmUuid, farmManageUuid, rowId)
}

// PostManageFarmsFarmUuidFarmManageUuidRowId operation middleware
func (siw *ServerInterfaceWrapper) PostManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "rowId" -------------
	var rowId int32

	err = runtime.BindStyledParameterWithOptions("simple", "rowId", c.Param("rowId"), &rowId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter rowId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostManageFarmsFarmUuidFarmManageUuidRowId(c, farmUuid, farmManageUuid, rowId)
}

// PutManageFarmsFarmUuidFarmManageUuidRowId operation middleware
func (siw *ServerInterfaceWrapper) PutManageFarmsFarmUuidFarmManageUuidRowId(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "rowId" -------------
	var rowId int32

	err = runtime.BindStyledParameterWithOptions("simple", "rowId", c.Param("rowId"), &rowId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter rowId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutManageFarmsFarmUuidFarmManageUuidRowId(c, farmUuid, farmManageUuid, rowId)
}

// PostManageFarmsFarmUuidFarmManageUuidRowIdCell operation middleware
func (siw *ServerInterfaceWrapper) PostManageFarmsFarmUuidFarmManageUuidRowIdCell(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "rowId" -------------
	var rowId int32

	err = runtime.BindStyledParameterWithOptions("simple", "rowId", c.Param("rowId"), &rowId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter rowId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostManageFarmsFarmUuidFarmManageUuidRowIdCell(c, farmUuid, farmManageUuid, rowId)
}

// PutManageFarmsFarmUuidFarmManageUuidRowIdCell operation middleware
func (siw *ServerInterfaceWrapper) PutManageFarmsFarmUuidFarmManageUuidRowIdCell(c *gin.Context) {

	var err error

	// ------------- Path parameter "farmUuid" -------------
	var farmUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmUuid", c.Param("farmUuid"), &farmUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "farmManageUuid" -------------
	var farmManageUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "farmManageUuid", c.Param("farmManageUuid"), &farmManageUuid, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter farmManageUuid: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "rowId" -------------
	var rowId int32

	err = runtime.BindStyledParameterWithOptions("simple", "rowId", c.Param("rowId"), &rowId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter rowId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PutManageFarmsFarmUuidFarmManageUuidRowIdCell(c, farmUuid, farmManageUuid, rowId)
}

// GetManagePlant operation middleware
func (siw *ServerInterfaceWrapper) GetManagePlant(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetManagePlant(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/groups/:groupUuid/manageFarms", wrapper.GetGroupsGroupUuidManageFarms)
	router.POST(options.BaseURL+"/groups/:groupUuid/manageFarms", wrapper.PostGroupsGroupUuidManageFarms)
	router.POST(options.BaseURL+"/login", wrapper.PostLogin)
	router.GET(options.BaseURL+"/manageFarms/:farmUuid", wrapper.GetManageFarmsFarmUuid)
	router.GET(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid", wrapper.GetManageFarmsFarmUuidFarmManageUuid)
	router.PUT(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid", wrapper.PutManageFarmsFarmUuidFarmManageUuid)
	router.GET(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid/:rowId", wrapper.GetManageFarmsFarmUuidFarmManageUuidRowId)
	router.POST(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid/:rowId", wrapper.PostManageFarmsFarmUuidFarmManageUuidRowId)
	router.PUT(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid/:rowId", wrapper.PutManageFarmsFarmUuidFarmManageUuidRowId)
	router.POST(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid/:rowId/cell", wrapper.PostManageFarmsFarmUuidFarmManageUuidRowIdCell)
	router.PUT(options.BaseURL+"/manageFarms/:farmUuid/:farmManageUuid/:rowId/cell", wrapper.PutManageFarmsFarmUuidFarmManageUuidRowIdCell)
	router.GET(options.BaseURL+"/managePlant", wrapper.GetManagePlant)
}