# openapi/bundled.yaml
openapi: 3.0.3
info:
  title: Wo-App API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /login:
    post:
      summary: Order to login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUser'
  /managePlant:
    get:
      summary: Get all users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
  /groups/{groupUuid}/manageFarms:
    get:
      summary: Get Farms by GroupUUID
      parameters:
        - name: groupUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FarmInfo'
    post:
      summary: Register New Farm
      parameters:
        - name: groupUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmInfoDetail'
      responses:
        '200':
          description: OK
  /manageFarms/{farmUuid}:
    get:
      summary: Get Farm Page Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FarmPageBasicInfo'
  /manageFarms/{farmUuid}/{farmManageUuid}:
    get:
      summary: Get furrow Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Farm page info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FurrowBasicInfo'
    put:
      summary: Update Farm FarmInfo
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarmInfoDetail'
      responses:
        '200':
          description: OK
  /manageFarms/{farmUuid}/{farmManageUuid}/{rowId}:
    get:
      summary: Get Furrow Cell Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rowId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Furrow Cell Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FurrowCellInfo'
    post:
      summary: Create Furrow Cell Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rowId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FurrowCellInfo'
      responses:
        '200':
          description: OK
    put:
      summary: Update Furrow Cell Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rowId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FurrowCellInfo'
      responses:
        '200':
          description: OK
  /manageFarms/{farmUuid}/{farmManageUuid}/{rowId}/cell:
    post:
      summary: Create Cell Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rowId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellInfo'
      responses:
        '200':
          description: OK
    put:
      summary: Update Cell Info
      parameters:
        - name: farmUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: farmManageUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rowId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellInfo'
      responses:
        '200':
          description: OK
components:
  schemas:
    LoginOrder:
      type: object
      properties:
        userId:
          type: string
        groupId:
          type: string
        password:
          type: string
          format: password
    LoginUser:
      type: object
      properties:
        userUuid:
          type: string
          format: uuid
        groupUuid:
          type: string
          format: uuid
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
    FarmInfo:
      type: object
      required:
        - farmName
        - farmUuid
      properties:
        farmName:
          type: string
        farmUuid:
          type: string
          format: uuid
    FarmInfoDetail:
      type: object
      required:
        - farmName
        - farmYear
        - farmSeasons
        - farmLength
        - farmWidth
      properties:
        farmName:
          type: string
        farmYear:
          type: integer
        farmSeasons:
          type: array
          items:
            type: integer
        farmLength:
          type: string
        farmWidth:
          type: string
        furrowNumber:
          type: integer
        furrowWidth:
          type: string
    FarmPageBasicInfo:
      type: object
      required:
        - farmName
        - farmYear
        - farmSeasons
        - farmManageUuid
      properties:
        farmName:
          type: string
        farmYear:
          type: integer
          format: int32
        farmSeasons:
          type: array
          items:
            type: integer
            format: int32
        farmManageUuid:
          type: string
          format: uuid
    FurrowBasicInfo:
      type: object
      required:
        - furrowNumber
      properties:
        furrowNumber:
          type: integer
          format: int32
    CellInfo:
      type: object
      required:
        - cellRow
        - cellColumn
        - cropItem
        - variety
        - datePlanted
        - growthStage
        - status
        - poorGrowthReason
        - dateHarvestPlanted
        - memo
      properties:
        cellRow:
          type: integer
          format: int32
        cellColumn:
          type: integer
          format: int32
        cropItem:
          type: string
        variety:
          type: string
        datePlanted:
          type: string
        growthStage:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        poorGrowthReason:
          type: integer
          format: int32
        dateHarvestPlanted:
          type: string
        memo:
          type: string
    FurrowCellInfo:
      type: object
      required:
        - furrowWidth
        - furrowLength
        - minPlantSpacing
        - rows
        - CellInfoArray
      properties:
        furrowWidth:
          type: integer
          format: int32
        furrowLength:
          type: integer
          format: int32
        minPlantSpacing:
          type: integer
          format: int32
        rows:
          type: integer
          format: int32
        CellInfoArray:
          type: array
          items:
            $ref: '#/components/schemas/CellInfo'